Move framework support to plugins
